added sprites
