+++ app/game/animation/test.html
@@ -4,6 +4,6 @@
     </head>
     <body>
         <canvas id="idleAnimation"></canvas>
-        <script src="test.js"></script>
+        <script src="animate.js"></script>
     </body>
 </html>
